fix(dilesa-proyectos-checklist-inline): hotfix 2 — sacar const/type de file 'use server'
ROOT CAUSE FOUND. El bug "x.map is not a function" se manifestaba en
la LISTA de anteproyectos (no solo en el detalle como pensé en el
hotfix 1). Stack: 'use server' files solo pueden exportar funciones
async; cuando se exporta un const (TAREA_ESTADOS_VALIDOS) o type
desde uno, Next.js queda `undefined` en cliente — cualquier `.map()`
sobre eso revienta.

Como <AnteproyectosModule> (la lista) importa `deriveAnalisis` de
anteproyecto-detalle.tsx, que importa <TareasChecklist>, que importa
TAREA_ESTADOS_VALIDOS de actions.ts, el bundle entero arrastra el
bug — incluso la página lista falla en el primer render porque
TAREA_ESTADOS_VALIDOS llega undefined cuando webpack resuelve el
chunk del checklist.

Fix:
- Crear components/dilesa/tareas-checklist-types.ts con
  TAREA_ESTADOS_VALIDOS + TareaEstado (archivo cliente normal, sin
  'use server').
- actions.ts importa de ahí (servidor también lo necesita para
  whitelist en updateTareaEstado).
- tareas-checklist.tsx importa de ahí.

Cero cambio de modelo. Cero cambio de comportamiento. Solo organización
de archivos para que Next.js no rompa el bundle.

También explica el React #418 (hydration mismatch) que veía Beto
PRIMERO en consola: el chunk del módulo se cargaba parcialmente,
el árbol de render entraba en estado inconsistente y la hidratación
fallaba antes de que el x.map siquiera ejecutara.
